[FEATURE] Add returnUrl in module "Links" (#86)
In the backend module "Example module" in the function "Links",
there are several links to edit records. After the "close"
button is clicked in the form, the expected behaviour would be,
to jump back to the previously opened module.

This functionality can be implemented, by setting returnUrl to
the module route. This is now implemented in one of the links
and an appropriate label added.

Resolves: #84