[TASK] Adjust embedding of user TSconfig
With #e78aaee17 the Configuration/TSconfig folder was renamed
to Configuration/TsConfig. Now, the path is corrected when
importing user TSconfig. Also the extension of the imported file
was adjusted to tsconfig.

Additionally, a superfluous import of page TSconfig was removed.
It was previously also available in ext_localconf.php before commit
#e78aaee17.