[FEATURE] Add example for sending HTTP request with Guzzle
This example will be used as code snippet for
https://docs.typo3.org/m/typo3/reference-coreapi/main/en-us/ApiOverview/Http/Index.html#basic-usage

The main part is the CatFactRequester class which requests a cat fact
from a public API. To hold the example small only the needed code is
available in this class. To really work with the CatFactRequester a
CatFactCommand is introduced which can be called with
`vendor/bin/typo3 examples:catfact` and displays a random fact
about cats.

Resolves: #108