[FEATURE] Improve the record link browser example
Isolate the record link TSconfig to its own file as it is easier to display and link to in the documentation. Move the TSconfig from "Extension" as the link browser is part of typo3/cms-backend and an integral part of the TYPO3 Core.

Adjust the TypoScript so the Haiku Detail view links to the detail view
of our non-Extbase Plugin example.

On the fly: Add comment-editor comments to the Haiku Plugin constants.

The custom github linkhandler TSconfig will be moved in a follow-up