[TASK] Streamline Haiku DetailController
With PHP 8.1 a new return type "never" was introduced. This type declares
that the method would - well - never return anything. So we add it to
notFoundAction(). Then we can streamline the main() method to return immediately
instead of assigning the result to a variable and return later.

This is a preparation for applying more Rector rules.