[TASK] Apply UP_TO_PHP_74 level set list to Rector rules (#158)
* [TASK] Apply UP_TO_PHP_74 level set list to Rector rules

This is the first part of applying the UP_TO_PHP_xx level set lists
until PHP 8.1. This way the changes are small.

The TernaryToNullCoalescingRector is part of the PHP 7.0 rule set
and can therefore be omitted.

The GithubLinkBuilder::TYPE_GITHUB constant was set to private manually
as the Rector rule would set it to public (which is not needed in this case).

The AddLiteralSeparatorToNumberRector is skipped as this would format exception codes
(timestamps) like 1_234_567_890 (which is not needed for readability here).