[TASK] Unify web-dir in composer.json
In order to unify some paths across extension, it is recommended
to use web-dir ".Build/Web".

For examples, this does not have any impact, as the web-dir is
currently not used.

But, in general, the paths are sometimes used in scripts (runTests.sh)
or configuration files (e.g. phpstan.neon), so it helps to unify things
if the common defaults are used.

This concerns:

- web-dir: .Build/Web
- bin-dir: .Build/bin
- vendor-dir: .Build/vendor

Related: TYPO3-Documentation/tea#802