[TASK] Add static code analyser `PHPStan` as code quality check
The TYPO3 Core and the Best Practice Team recommends
to have a static code analyzer in the defense line,
and preferes PHPStan due to experience with that tool.

This change now adds the static code analyser `PHPStan`
with a first configurattion and a lower level as a start.
Levels can be raised step by step in follow-ups.

Note: XClass `Classes/Xclass/NewRecordController.php`
example does not work anymore, most likely since v12.
Added a comment to emphasize this, and that it should
be reworked or removed.

PHPStan can be run with

```terminal
Build/Scripts/runTests.sh -s phpstan
```

Generate baseline with

```terminal
Build/Scripts/runTests.sh -s phpstanBaseline
```